requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
sqlalchemy==2.0.39
pydantic-settings==2.8.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
